+++ CHANGES.rst
@@ -6,10 +6,14 @@ There's a frood who really knows where his towel is.
 1.0a2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Remove browser layer and JS resource registry registrations.
+  Include all Cycle2 and plugins code as browser resources (closes `#1`_).
+  [rodfersou]
 
 
 1.0a1 (2014-10-20)
 ------------------
 
 - Initial release.
+
+.. _`#1`: https://github.com/collective/collective.js.cycle2/issues/1